separate_email_cleaner_process/requirements.txt
asyncio==3.9.1
httpx==0.25.2
msgraph-sdk-python==1.8.0
requests==2.31.0
requests-oauthlib==1.3.1
typing-extensions==4.9.0